+++ physics/src/calibration/pads/baseline/tests.rs
@@ -109,13 +109,3 @@ fn try_pad_baseline_correctness_sim() {
         }
     }
 }
-
-#[test]
-#[should_panic]
-fn safe_guard_try_pad_baseline() {
-    let column = TpcPadColumn::try_from(0).unwrap();
-    let row = TpcPadRow::try_from(0).unwrap();
-    let pad_position = TpcPadPosition { row, column };
-
-    let _ = try_pad_baseline(u32::MAX - 1, pad_position);
-}